Move JPA DAO implementations to separate module
