+++ library/src/main/java/ua/nykyforov/service/library/application/dao/jpa/JpaAuthorDao.java
@@ -1,4 +1,4 @@
-package ua.nykyforov.service.library.application.dao;
+package ua.nykyforov.service.library.application.dao.jpa;
 
 import org.springframework.stereotype.Repository;
 import ua.nykyforov.service.library.core.dao.AuthorDao;
