+++ library/src/main/java/ua/nykyforov/service/library/application/dao/jpa/JpaBookDao.java
@@ -1,4 +1,4 @@
-package ua.nykyforov.service.library.application.dao;
+package ua.nykyforov.service.library.application.dao.jpa;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
